% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_plot.R
\name{attrition_cycle_plot_int}
\alias{attrition_cycle_plot_int}
\title{Attrition Plot Interactive}
\usage{
attrition_cycle_plot_int(dfCycle, l.category = NULL, ExitPath = "TOTAL")
}
\arguments{
\item{ExitPath}{a character string indicating which exit path ("TOTAL", "With Masters", "Without Masters"
to plot )}
}
\value{
A cycle plot attrtition data as a plotly object
}
\description{
\code{attrition_cycle_plot_int} plotting funciton for the creation of cycle plots For an overivew of cycle plots,
please see this \href{https://www.perceptualedge.com/articles/guests/intro_to_cycle_plots.pdf}{article} in
Stephen Few's Perceptual Edge library.
}
