% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{filter.ir}
\alias{filter.ir}
\title{Subset rows in \code{ir} objects using column values}
\source{
\code{\link[dplyr:filter]{dplyr::filter()}}
}
\usage{
filter.ir(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{An object of class \code{ir}.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Expressions that return a
logical value, and are defined in terms of the variables in \code{.data}.
If multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\description{
Subset rows in \code{ir} objects using column values
}
\examples{
## filter
dplyr::filter(ir_sample_data, sample_type == "office paper")


}
