% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dir.R
\name{get_data_dir}
\alias{get_data_dir}
\title{get_data_dir}
\usage{
get_data_dir(folder_path = NULL)
}
\description{
When working in the URI sharepoint structure
function will return the path to the data director
}
