% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-package.R
\docType{package}
\name{ir-package}
\alias{ir}
\alias{ir-package}
\title{ir: Functions to Handle and Preprocess Infrared Spectra}
\description{
Functions to import and handle infrared spectra (import from '.csv' and Thermo Galactic's '.spc', baseline correction, binning, clipping, interpolating, smoothing, averaging, adding, subtracting, dividing, multiplying, atmospheric correction, 'tidyverse' methods, plotting).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://henningte.github.io/ir/}
  \item \url{https://github.com/henningte/ir/}
  \item Report bugs at \url{https://github.com/henningte/ir/issues/}
}

}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID}) [copyright holder]

}
\keyword{internal}
