% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_check_ir}
\alias{ir_check_ir}
\title{Checks if an object is of class \code{ir}}
\usage{
ir_check_ir(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[ir:ir_new_ir]{ir}}.
}
\description{
\code{ir_check_ir} checks if an object is of class \code{\link[ir:ir_new_ir]{ir}}.
}
