% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_check_metadata}
\alias{ir_check_metadata}
\title{Checks infrared spectra metadata.}
\usage{
ir_check_metadata(x)
}
\arguments{
\item{x}{A \code{data.frame} with a column
\code{sample_id} and a column \code{measurement_id} and optional
additional columns containing
metadata for the samples. Optionally, an empty \code{data.frame}
can be defined if no metadata are available.}
}
\value{
A \code{data.frame} that matches the requirements of the
argument \code{metadata} of \code{\link{ir_new_ir}}.
}
\description{
\code{ir_check_metadata} checks if a \code{data.frame}
matches the requirement of the argument \code{metadata} of
\code{\link{ir_new_ir}}.
}
