% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_check_spectra}
\alias{ir_check_spectra}
\title{Checks a list of spectra}
\usage{
ir_check_spectra(x)
}
\arguments{
\item{x}{A named list in which each element contains spectral data
for one measurement. Each list element
must be a \code{data.frame} with two columns and a row for each
wavenumber value in the spectra data. The first column must contain
unique wavenumber values and the second column intensity values of
the measured spectrum of the sample.}
}
\value{
A list that matches the requirements of the argument \code{spectra}
of \code{\link{ir_new_ir}}.
}
\description{
\code{ir_check_spectra} checks if a list of infrared spectra
matches the requirement of the argument \code{spectra} of
\code{\link{ir_new_ir}}.
}
