% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_clip.R
\name{ir_clip}
\alias{ir_clip}
\title{Clips infrared spectra to new wavenumber ranges.}
\usage{
ir_clip(x, range)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{range}{A numeric vector of length two
with the first element specifying the start
of the new wavenumber range and the second
element specifying the end of the new wavenumber
range.}
}
\value{
An object of class \code{ir}.
}
\description{
\code{ir_clip} clips infrared spectra to a new, specified,
wavenumber range.
}
