% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_drop_unneccesary_cols}
\alias{ir_drop_unneccesary_cols}
\title{Drop all non-required columns in an ir object}
\usage{
ir_drop_unneccesary_cols(x)
}
\arguments{
\item{x}{An object of class \code{\link{ir}}}
}
\description{
Drop all non-required columns in an ir object
}
\examples{
x1 <-
   ir::ir_sample_data \%>\%
   ir_drop_unneccesary_cols()
}
\keyword{internal}
