% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_flat.R
\name{ir_flat_check}
\alias{ir_flat_check}
\title{Checks if an object is of class \code{ir_flat}}
\usage{
ir_flat_check(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class \code{\link[=ir_new_ir_flat]{ir_flat()}}.
}
\description{
\code{ir_flat_check} checks if an object is of class \code{\link[=ir_new_ir_flat]{ir_flat()}}.
}
