% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_import_csv.R
\name{ir_import_csv}
\alias{ir_import_csv}
\title{Imports infrared spectra from various files.}
\usage{
ir_import_csv(filenames, sample_id = NULL, ...)
}
\arguments{
\item{filenames}{A character vector representing the complete paths to
the .csv files to import.}

\item{sample_id}{A character vector with the same length as \code{filenames}
representing the name of the sample. If not provided (i.e. \code{sample_id = NULL}),
the sample name will be extracted from \code{filenames}.}

\item{...}{Further arguments passed to \code{\link[utils:read.table]{read.csv}}.}
}
\value{
An object of class \code{\link[ir:ir_new_ir]{ir}} containing the infrared spectra
extracted from the .csv file(s).
}
\description{
\code{ir_import_csv} imports raw infrared spectra from a .csv file
that contains one spectrum, with x axis values (e.g. wavenumbers) in
the first column and intensity values in the second column, or several
of such files. Note that
the function does not perform any checks for the validity of the
content read from the .csv file.
}
