% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_flat.R
\name{ir_new_ir_flat}
\alias{ir_new_ir_flat}
\title{Creates an object of class \code{ir_flat}}
\usage{
ir_new_ir_flat(x)
}
\arguments{
\item{x}{A \code{data.frame} with only numeric columns and only the first
column name being "x".}
}
\value{
An object of class \code{ir_flat}.
}
\description{
\code{ir_new_ir_flat} is the constructor function for objects of class \code{ir_flat}.
An object of class \code{ir_flat} is a \code{data.frame} where the first
column (\code{"x"}) contains unique x values of spectra (e.g. wavenumbers)
and all remaining columns represent intensity values from spectra
corresponding to the x values.
}
\examples{
x_flat <-
   ir::ir_sample_data \%>\%
   ir::ir_flatten()

}
