% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_normalise.R
\name{ir_normalise}
\alias{ir_normalise}
\title{Normalises infrared spectra.}
\usage{
ir_normalise(x, method = "area")
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{method}{A character value specifying which normalisation method to apply.
If \code{method = "zeroone"}, all intensity values will be normalised to [0;1].
If \code{method = "area"}, all intensity values will be divided by the sum of
the intensity values at all wavenumber values of the spectrum. If
\code{method} is convertable to a numeric value, e.g. \code{method = "980"},
the intensity of all spectra at a wavenumber value of 980 will be set to 1 and
the minimum intensity value of each spectrum will be set to 0, i.e. the spectra
will be normalised refering to a specific wavenumber value.}
}
\value{
An object of class \code{ir} representing a normalised version of \code{x}.
}
\description{
\code{ir_normalise} normalises the intensity values of infrared spectra.
Spectra can be normalised in three ways:
\enumerate{
  \item Normalisation so that the intensity values range in [0;1]. Note that
  for different spectra, for different wavenumber values the intensity may be
  1 after normalisation, depending on the location of the peak with the maximum
  height.
  \item Normalisation so that the area under the spectral curve sums to 1. Note
  that in the case of negative intensities values, these will be count as negative
  values during summation.
  \item Normalisation so that the intensity at a specified wavenumber value
  has value 1 and the minimum intensitiy value is 0.
}
}
