% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_remove_missing.R
\name{ir_remove_missing}
\alias{ir_remove_missing}
\title{Removes empty data values in objects of class \code{ir}.}
\usage{
ir_remove_missing(x, remove_rows = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{remove_rows}{A logical value indicating if rows in \code{x}
with empty spectra should be discarded (\code{remove_rows = TRUE})
or not (\code{remove_rows = FALSE}).}
}
\value{
\code{x} with cleaned spectra.
}
\description{
\code{ir_remove_missing} takes and object of class \code{ir} and removes
all rows in the \code{data.frame}s of the list column \code{spectra} that
have \code{NA} intensity values (column \code{y}). Additionally, one
can specify to remove rows in the \code{ir} object to discard if they
contain empty spectra.
}
