% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_snv.R
\name{ir_snv}
\alias{ir_snv}
\title{Standard Normal Variate correction for \code{ir} objects}
\source{
Code adapted from the \href{https://cran.r-project.org/web/packages/prospectr}{'prospectr'} package.
}
\usage{
ir_snv(x, return_ir_flat = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[=ir_new_ir]{ir}} containing the spectra to
correct.}

\item{return_ir_flat}{Logical value. If \code{TRUE}, the spectra are returned as
\code{\link[=ir_new_ir_flat]{ir_flat}} object.}
}
\value{
\code{x} after Standard Normal Variate correction (or the same as
\code{ir_flat} object if \code{return_ir_flat = TRUE}).
}
\description{
Standard Normal Variate correction for \code{ir} objects
}
\examples{
x1 <-
  ir_sample_data \%>\%
  ir_snv()

}
