% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_subtract.R
\name{ir_subtract}
\alias{ir_subtract}
\title{Subtract infrared spectra}
\usage{
ir_subtract(x, y)
}
\arguments{
\item{x}{An object of class \code{\link{ir}}.}

\item{y}{An object of class \code{\link{ir}} with the same number of rows as
\code{x} and the same x axis values (e.g. wavenumber values) in each matching
spectrum as in \code{x}.}
}
\value{
\code{x} where for each spectrum the respective intensity values in
\code{y} are subtracted.
}
\description{
\code{ir_subtract} takes two objects of class \code{ir}, \code{x} and
\code{y}, and subtracts the intensity values of spectra in matching rows from
\code{y} from that of \code{x}.
}
\examples{
x1 <-
  ir::ir_subtract(ir::ir_sample_data, ir::ir_sample_data)
}
