% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\name{Ops.ir}
\alias{Ops.ir}
\title{Arithmetic operations for \code{ir} objects}
\usage{
\method{Ops}{ir}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ir}.}

\item{e2}{An object of class \code{ir} or a numeric value.}
}
\description{
Arithmetic operations for \code{ir} objects
}
\examples{
## addition
ir::ir_sample_data + ir::ir_sample_data
ir::ir_sample_data + 2

## subtraction
ir::ir_sample_data - ir::ir_sample_data
ir::ir_sample_data - 2

## multiplication
ir::ir_sample_data * ir::ir_sample_data
ir::ir_sample_data * 2

## division
ir::ir_sample_data / ir::ir_sample_data
ir::ir_sample_data / 2

}
