% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ir.R
\name{plot.ir}
\alias{plot.ir}
\title{Plots an object of class ir}
\usage{
\method{plot}{ir}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=ir_new_ir]{ir}}.}

\item{...}{Further arguments, will be ignored.}
}
\value{
An object of class \code{\link[ggplot2:ggplot]{ggplot2()}}.
}
\description{
\code{plot.ir} is the plot method for objects of class \code{ir}.
}
\examples{
# simple plotting
plot(ir::ir_sample_data[1:2, ])

# advanced functions
plot(ir::ir_sample_data) +
   ggplot2::facet_wrap(~ sample_type)
}
