% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FRETF.R
\name{read_FRETF}
\alias{read_FRETF}
\title{read_FRETF}
\usage{
read_FRETF(file, dir = NA)
}
\description{
Function for loading the Freshman Cohorts files (FRETF)
flat file (.dat) from e-Campus.
}
\details{
The Freshman Cohort file is used for reporting retention
and graduation rate

To retireve the dat file:

1.	Login to eCampus.
2.	Select:  URI Interfaces / Institutional Reporting / Process / Student File Download.
}
