% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GRADyyyy.R
\name{read_GRADyyyy}
\alias{read_GRADyyyy}
\title{read_GRADyyyy}
\usage{
read_GRADyyyy(file, dir = NA)
}
\description{
Function for loading the Graduation lists
flat file (.dat) from e-Campus.
}
\details{
To retireve the dat file:
}
