% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_internal_payroll.R
\name{read_internal_payroll}
\alias{read_internal_payroll}
\title{read_internal_payroll}
\usage{
read_internal_payroll(file, data_dir = NULL)
}
\description{
Internal payroll report uses a non-standard data format
this function loads and reformats the data to facilitate
analysis
}
