% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PSW01.R
\name{read_PSW01}
\alias{read_PSW01}
\title{read_PSW01}
\usage{
read_PSW01(file, dir = NA)
}
\description{
Function for loading the Student Enrollment History
flat file (.dat) from e-Campus.
}
\details{
The Student Enrollment History file is used for Final Enrollment
Reports and other reports involving headcount, age distribution,
credits, etc.

To retireve the dat file:

1.	Login to eCampus.
2.	Select:  URI Interfaces / Institutional Reporting / Process / Student File Download.
3.	Run the process after entering a TERM (cyym) at the prompt.
4.	At completion, run FTP to transfer the dataset to the PC (translate to ASCII format).
5.	Rename as  PSW01-tyy-mmmdd.DAT  where t is term (F, S, or SUM), yy are the term year
digits, mmm is the month of the extract, and dd is the day of the extract.  Example:
PSW01-F03-OCT15.DAT.
6.	Clean up any erroneous/missing data and save a new copy.
}
