% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_STDNT_ENRL.R
\name{read_STDNT_ENRL}
\alias{read_STDNT_ENRL}
\title{read_STDNT_ENRL}
\usage{
read_STDNT_ENRL(file, data.dir = NA)
}
\description{
Function for loading the Course Enrollment Files
flat file (.dat) from e-Campus.
}
\details{
files are located: Data/Raw/Course Enrollment/
}
