% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{rep.ir}
\alias{rep.ir}
\title{Replicate ir objects}
\usage{
\method{rep}{ir}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=ir_new_ir]{ir}}.}

\item{...}{See \code{\link[=rep]{rep()}}.}
}
\value{
An object of class \code{ir} with replicated spectra.
}
\description{
\code{rep.ir} is the replicate method for \code{\link[=ir_new_ir]{ir}} objects.
Replicating and \code{ir} object means to replicate its rows and bind these
together to a larger \code{ir} object.
}
\examples{
# replicate the sample data
x <- rep(ir::ir_sample_data, times = 2)
x <- rep(ir::ir_sample_data, each = 2)
x <- rep(ir::ir_sample_data, length.out = 3)

}
