% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{select.ir}
\alias{select.ir}
\title{Subset columns in \code{ir} objects using column names and types}
\source{
\code{\link[dplyr:select]{dplyr::select()}}
}
\usage{
select.ir(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{ir}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\description{
Subset columns in \code{ir} objects using column names and types
}
\examples{
## select
dplyr::select(ir_sample_data, spectra)
dplyr::select(ir_sample_data, holocellulose) # drops ir class

}
