% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{separate_rows.ir}
\alias{separate_rows.ir}
\title{Separate a collapsed column in an \code{ir} object into multiple rows}
\source{
\code{\link[tidyr:separate_rows]{tidyr::separate_rows()}}
}
\usage{
separate_rows.ir(data, ..., sep = "[^[:alnum:].]+", convert = FALSE)
}
\arguments{
\item{data}{An object of class \code{ir}.}

\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to separate across
multiple rows}

\item{sep}{Separator delimiting collapsed values.}

\item{convert}{If \code{TRUE} will automatically run
\code{\link[=type.convert]{type.convert()}} on the key column. This is useful if the column
types are actually numeric, integer, or logical.}
}
\description{
Separate a collapsed column in an \code{ir} object into multiple rows
}
\examples{
## separate_rows
ir_sample_data \%>\%
  tidyr::unite(
    col = content, holocellulose, klason_lignin
  ) \%>\%
  tidyr::separate_rows(
    col
  )


}
