% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-package.R
\docType{package}
\name{ir-package}
\alias{ir}
\alias{ir-package}
\title{ir: Functions to Handle and Preprocess Infrared Spectra}
\description{
Functions to import and handle infrared spectra (import from .csv and Thermo Galactic's .spc, baseline correction, binning, clipping, interpolating, smoothing, averaging, adding, subtracting, dividing, multiplying, plotting).
}
\author{
\strong{Maintainer}: Henning Teickner \email{h_teic01@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID})

}
\keyword{internal}
