% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_bc.R
\name{ir_bc}
\alias{ir_bc}
\title{Performs baseline correction on infrared spectra.}
\usage{
ir_bc(x, method = "rubberband", degree = 2, return_bl = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{method}{A character value indicating which method should be used
for baseline correction. If \code{method = "polynomial"}, a polynomial
is used for baseline correction. If \code{method = "rubberband"}, a
rubberband function is used for baseline correction.}

\item{degree}{An integer value representing the degree of the polynomial
used for baseline correction if \code{method = "polynomial"}.}

\item{return_bl}{A logical value indicating if for each spectrum the baseline
should be returned in addition to the corrected intensity values
(\code{return_bl = TRUE}) or not (\code{return_bl = FALSE}).}
}
\value{
An object of class \code{ir} with the baseline
corrected spectra and if \code{returnbl = TRUE} a new list column
"baselines" with the baselines.
}
\description{
\code{ir_bc} performs baseline correction for infrared
spectra. Baseline correction
is either performed by using a polynomial with user defined
degree fitted to each spectrum (see
\code{\link[ChemoSpec:baselineSpectra]{baselineSpectra}}) or by using a rubberband
function that is fitted to each spectrum
(see \code{\link[hyperSpec:spc.rubberband]{spc.rubberband}}).
}
\seealso{
\code{\link{ir_bc_polynomial}},
\code{\link{ir_bc_rubberband}}.
}
