% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_interpolate_region.R
\name{ir_interpolate_region}
\alias{ir_interpolate_region}
\title{Interpolates selected regions in infrared spectra.}
\usage{
ir_interpolate_region(x, region)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{region}{A numerical vector of length two with the first value
representing the start (smaller) wavenumber value of the region and the
second value representing the end (larger) wavenumber value of the region
to interpolate.}
}
\value{
\code{x} with the defined wavenumber region interpolated linearly.
}
\description{
\code{ir_interpolate_region} linearly interpolates a user-defined
region in infrared spectra.
}
