% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{rowwise.ir}
\alias{rowwise.ir}
\title{Group input \code{ir} objects by rows}
\source{
\code{\link[dplyr:rowwise]{dplyr::rowwise()}}
}
\usage{
rowwise.ir(.data, ...)
}
\arguments{
\item{.data}{Input data frame.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variables to be preserved
when calling \code{\link[dplyr:summarise]{summarise()}}. This is typically a set of variables whose
combination uniquely identify each row.

\strong{NB}: unlike \code{group_by()} you can not create new variables here but
instead you can select multiple variables with (e.g.) \code{everything()}.}

\item{data}{An object of class \code{ir}.}
}
\description{
Group input \code{ir} objects by rows
}
\examples{
## rowwise
dplyr::rowwise(ir_sample_data) \%>\%
  dplyr::mutate(hkl = mean(klason_lignin, holocellulose))

}
